move examples into separate folder
